Rewrite rar comment detection
So far, the detection only worked (partly) for rar/unrar version 4,
but not for earlier versions or recent version 5 of the command line
tools. Now, all these versions (at least starting with version 3) are
supported.

A failed detection could make xarchiver hang or even crash.

Because comments in rar archives are difficult to extract, the detection
was never foolproof and it still isn't, but it is much safer now.