Restore visible directory path selection upon notebook page switch
When working with more than one archive and with selection of a
directory in an archive's directory tree view, all the content will be
redisplayed when switching from another page, but the visible selection
of the directory gets lost.

Restoring it isn't that simple. A call to xa_sidepane_select_row() will
call gtk_tree_selection_select_iter() in _xa_sidepane_select_row() which
will trigger xa_sidepane_row_selected(). Since we are currently handling
a switch-page signal, the current notebook page is still the old one and
therefore xa_sidepane_row_selected() can't find the new archive's index.

Fortunately, all display is already set up and we'll just need the
visible selection marker (and the scroll to the cell), so we temporarily
block the signal that would trigger xa_sidepane_row_selected().