Redesign dialog to create a new archive
Automatically add the respective file extension for the selected archive
type.

The respective file extension is in ArchiveSuffix, so search for the
matching ArchiveSuffix belonging to ArchiveType.

In order for that to work, there must be an ArchiveType (even a dummy,
empty one) for every ArchiveSuffix (ArchiveSuffix must not be empty).

When selecting an other archive type from the combo box, change nothing
but the file extension (the already typed file name remains untouched).

Loosely based on a patch by Teo Box, tbox1 users.sf net.