Don't automatically convert to UTF-8
There's no point in converting to UTF-8 when removing the path
component from an archive name in xa_remove_path_from_archive_name().
The conversion to UTF-8 has to be done where UTF-8 encoding is actually
required (which is not always the case). Any input in local filename
encoding should provide output in local filename encoding as well.

As a result, UTF-8 encoding has to be performed at several places
instead.