Implement a new function to ask about (un)compressor capabilities
When opening an archive there is information about its capabilities,
e.g. whether it can delete added files or whether password protection
is available.

When creating a new (i.e. empty archive) there is no such information
at all. So implement a new function that can request the capabilities
of a compressor without having to call it (by opening anything). This
new function can be called even with a just initialized, bare XArchive.
The only thing needed to know at that point is the XArchiveType.

This will enable available options in the add dialog of a new archive
which formerly all were disabled.