Use but don't demand g_mkdtemp()
Since commit 9f8d0463229466321f482af738c5ee170e39cf10 we are using
GLib's mkdtemp(), but this unfortunately requires GLib 2.30 while
for the rest of the code GLib 2.12 is still sufficient.

So, in order to not unnecessarily raise requirements and to still be
okay with GLib 2.12 either use C function mkdtemp() (if available) or
a mkdtemp() implementation by ourselves to define g_mkdtemp(). Only
use real g_mkdtemp() if GLib 2.30 or newer happens to be available.

Additionally, adjust prototype declaration and return type of our
mkdtemp().