Sort messages in .po files
We have already removed the line numbers from location comments
so that changes are a lot smaller and clearer now.

But the position of a msgid in a .po file depends on its position
in the source file which means that code reorganization (without
any change of messages) would still cause unnecessary disorder in
the .po files. In order to eliminate that effect, sort the msgids
alphabetically by default.

This just reorders the msgids. No new messages have been added.