Add missing XA_ARCHIVESTATUS_IDLE and XA_ARCHIVESTATUS_ERROR
The status is currently the status of the last (un)compressor call,
i.e. it does not change after the call has been executed.

Change it now to XA_ARCHIVESTATUS_ERROR if the child program couldn't
be executed and to XA_ARCHIVESTATUS_IDLE when the child program ends
normally. (It already is XA_ARCHIVESTATUS_ERROR if the child program
ends with error.)

(This will momentarily break some code that depends on the old
behavior.)