Rebuild child related code
After having spawned a child asynchronously there are three processes
to be watched:

* the child's standard output
* the child's standard error
* the child's exit status

Post-processing (i.e. user interface actions) must not start until all
of the three processes have finished. So move all post-processing code
to the post-processing function (currently named xa_watch_child()).
Furthermore, the post-processing function must not assume that it is
called last, so add a new function xa_process_exit() and move all exit
related code to it.

Each process now reports to the post-processing function when it is
done. This will be tracked by new XArchive member child_ref.

This rebuild is first step and a rather dirty approach that aims for the
goals to not break executability and to make minimal, easy recognizable
changes. It will be enhanced one by one.