Improve unarj support
Unarj cannot extract individual files, it always extracts all files
from an archive, and it fails with error if a file already exists.
It is therefore impossible to extract twice (different files) from
an archive.

Overcome this by improving the extraction command. Extract to a
temporary directory first and subsequently move (only) the selected
files to their destination (if necessary).