Revise the context menu
Don't allow the open popup for a multiple selection and
don't allow renaming directory entries.

(Every compressor needs different options in order to handle the
renaming of a directory which currently is composed of extracting,
deleting and adding a directory, and a particular compressor can
even need different options for any of the three steps).