Relocate reversing the output to xa_child_processed()
The list has been built by prepending lines, so it must be
reversed at the end, but don't do so until all processes have
stopped.

This also reverses the error output which was missing so far.

(These are not the right and final places, but as long as
xa_child_processed() is not yet rewritten, it'll do.)