Completely rework the counting of columns (part 1)
The current counting and usage is confusing.

Rename structure XArchive member nc columns.

Every archive type has three fixed columns, namely the 1st one (pixbuf),
the 2nd one (filename) and the last one (an internally used pointer).
The remaining columns are specific to the archive type.

This means that a specific archive type has to define archive->columns
minus 3 individual items and titles in order to display the list of its
content. Struct XEntry member columns will later point to these very
archive->columns minus 3 columns.

So structure XArchive member columns is now exactly the number of
columns of the liststore/treemodel that holds the data for the list
of the archive content.

Additionally, rename names[] titles[], change the type to gchar * and
add const to types[] and titles[].