Completely rework the counting of columns (part 2)
The current counting and usage is confusing.

Utilize structure XArchive member columns which is the number of columns
of the liststore/treemodel that holds the data for the archive content
list.

So:

* Column 0 is a pixbuf for the icon.

* Column 1 is the filename.

* Columns 2 to archive->columns minus 2 are the archive's individual
  columns (items).

* Column archive->columns minus 1 is an internally used pointer to the
  XEntry holding all the data for the respective treeview line.

The treeview conflates columns 0 and 1 and the last column is invisible,
i.e. an archive's data is presented in archive->columns minus 2 columns
to the user.

This rework automatically fixes a bug with allocating unneeded space for
the filename in XEntry's columns.