Enable opening archives in the background
As long as no user interface function is called, it is safe to open an
archive and parse it in the background. This can be used in batch mode
in order to check whether entries are encrypted and the archive is
password protected.

The functions not to be called are xa_create_liststore() on opening and
xa_set_archive_entries_for_each_row() during parsing.

Allow calling the parse_output function without xa_main_window in order
to make that work.