Rewrite batch mode password detection xa_detect_encrypted_archive()
The function was limited to arj and zip and expended quite some effort
in order to detect password protection.

We already can easily detect password protection for every archive type
while parsing the archive's output. Utilize the recent possibility to
open archives in the background for xa_detect_encrypted_archive().

As a side effect, header encrypted archives will now be detected and
handled as well.