Rewrite xa_child_processed()
Collect the statuses of all three child processes but begin
post-processing not until all three child processes have finished.

If all three child processes exited normally and we aren't in batch
mode, reverse the output (which is in reverse order for performance
issues), possibly show archive comment and/or command output and
update the user interface with the archive's parsers output and its
properties and capabilities.
Finally and in any case, put an end to archive output parsing and
set the archive's status to "idle" (meaning everything's okay).

If any of the three child processes exited with a return code indicating
an error, show the command's output if requested and set the archive's
status to "error".