Rewrite xa_run_command() (part 2)
Pass the command as a simple string rather than in a GSList.

xa_run_command() is called in more than 90% of cases with a list
containing just one single element. Only four calls use the list
with two or more commands. It is extreme overhead to always create
a GSList for that (and to clean it up afterwards).

In the rare cases where more commands must be executed, several
consecutive calls can be made just as well.

Additionally, remove unnecessary initializations.