Remove all compressed tar related code from gzip_et_al
In order to disentangle the code of gzip_et_al and tar concerning
compressed tar archives and to get rid of any filename extension
related decision of the archive type in the near future, the plan
behind the ongoing code reorganization of gzip_et_al and tar is
the following:

A compressed tar archive is ordinary compressed data (by currently
bzip2, gzip, lzma, lzop or xz) like any other data compressed by
one of these. So the compressor shall be determined and the data
shall be passed to the gzip_et_al open function xa_gzip_et_al_open().

xa_gzip_et_al_open() shall decompress the passed data now and check
whether the content is a tar archive. If not, xa_gzip_et_al_open()
shall continue as usual. If so, it shall change the archive type to
the compressed tar type, reset the already set archive capabilities,
set the compressed tar type capabilities instead, change the archive
functions to those for a normal tar archive and finally call tar's
open function. All processing of the compressed tar archive shall be
performed by tar and shall be coded in tar.c.

So the only tar related code in gzip_et_al shall be the check for a
tar archive and the preparation of the XArchive structure similar to
parts of the initialization by xa_init_archive_structure().

(This commit makes gzip_et_al.c compile again.)