Enable adding for gzip_et_al type archives
Allow to add only if a gzip_et_al type archive is new and empty
(i.e. created by Archive -> New). As soon as it is opened or
reloaded the first time, disabled adding.

Use revised code from xa_tar_add() for new xa_gzip_et_al_add().

(This is part of the effort to disentangle the code of gzip_et_al
and tar.)