Rewrite xa_tar_add()
If it's a new and empty archive (i.e. created by Archive -> New) either
directly create the requested uncompressed tar archive or, in case a
compressed tar archive has been requested, prepare a temporary working
copy representing the decompressed version of the archive.

From now on, the archive will exist.

Operate on the decompressed tar working copy in the first place. If it
is a compressed tar archive, compress the result afterwards.