Simplify xa_gzip_et_al_open()
While most single file uncompressors are restrictive and refuse to
work otherwise, there are uncompressors that don't need to have a
specific file name suffix. In this case they choose a file name for
the uncompressed file by themselves which made it necessary to read
the output directory in order to get the file name.

Now, with lz4, there is an other issue: lz4 doesn't create an output
file but sends results to standard output if no destination name is
provided.

To overcome both issues, we always uncompress to standard output and
redirect output to a file the name of which we know.