Rewrite detection of a rar archive version 5
Relocate the detection from xa_rar5_parse_output() to
xa_detect_archive_type().

The former detection in xa_rar5_parse_output() is too late
for (un)compressors which need to know the exact archive version
prior to operate on it. What's more, we already have that information
at archive type analyzing time, in xa_detect_archive_type().
Unfortunately, there isn't yet an allocated XArchive structure to store
the version information in, so we have to return both type and version
in a new struct ArchiveType.

This also fixes an issue with uncompressors which can handle rar
archives, but not version 5 ones, and which do not return an error
code but only a console message like "unsupported archive format".
There was nothing displayed and nothing reported by Xarchiver in
that case.