Remove column 'GUA'
This is indicating the set group ID, set user ID and the restricted
deletion flag or sticky bit, but due to a bug in open-source ARJ up to
and including version 3.10.22 it never worked as designed and only
showed three dashes.

Next version of open-source ARJ will show this information in the
column 'Attributes' in the usual Unix-like style.