Fix bug with wrong information displayed for some directories
Directories which either don't have an own entry in an archive contents
list or the entry of which appears after an entry of a file with a path
containing that directory don't show their own column information, but
rather the information of the file in the path of which they appear
first.

Fix it by not filling the archive entry columns for every new entry,
because there is not yet any information for directory entries that
must be created "on the fly" (because they appear in a file path).

Fill the archive entry columns for every file or directory with an own
entry in the archive contents list, no matter where in the contents list
they appear (i.e. no matter whether the XEntry already exists or not).

Directories without an own entry in an archive contents list will
consequently lack any column information, because it simply isn't
available.