Improve the generic column parser
Try to detect and handle filenames which contain leading whitespace.

The detection is based upon the assumption that the column for the
filenames is fixed (i.e. the same column where the first filename
without leading whitespace has been found). If the first filename
contains leading whitespace, it will be displayed without its leading
whitespace (and thus cannot be extracted).