Improve recognition of pre-POSIX tar archives
These archives have no magic number in the header, but simply checking
for a couple of zeros is too vague, because they may easily get confused
with other data being filled with zeros after a small header or at the
beginning.

Check at least that the next block starts with either a file name or
some data.