Fix the generic column parser
The initialization (USE_PARSER) added with commit
026c74cf9dbb9e69f03bc9b85995362c4e07b81b is insufficient.

There is only a single static variable to save the position
where a filename has been found, but the same parse_output
function may be used for different archive types, thus the
next use of the parser will start with that very and wrong
position information.

So, when the archive changes, reinitialize the position variable.