Block "drag-data-get" while the child process is running
This enables both the flashing LED status and the stop button,
indicating that xarchiver is busy.

Instead of showing a message dialog (which isn't properly usable
during a drag action), use a stop icon for the drag. This icon is
more conspicuous, and the drag can easily be released.

This partly fixes github issue #41, reported by mirh.