Add capability to extract without full path for cpio style archives
Cpio style archives don't provide an option to extract without full
path. In order to implement such extraction this has to be coded
manually.

So we extract with full path in a temporary directory first and collect
all extracted files by traversing the temporary directory then. In a
final step, we move these files to the archive's extraction directory.

Extraction without full path is necessary for operations from inside the
archive when activating a file's context menu.