Remove weird and unnecessary condition
xa_main_window is a GtkWidget pointer, so the test for FALSE is
nonsensical.

Moreover, due to program code changes in the meantime, the whole
second condition has become unnecessary, because the check whether
an archive has a password is now performed even before the dialog
appears (and this with or without main window).