Do not unconditionally change the environment variable DISPLAY
DISPLAY is set before calling external programs with graphical user
interface, so that these programs run on the same display as xarchiver
itself (which could run on a different display than the one it was
started from with option --display).

However, if xarchiver is running from inside a Wayland compositor,
DISPLAY is already set and would otherwise be overwritten with the
Wayland display.

This fixes github issue #96, reported by Exagone313.