Fix gzip_et_al's add command
In commit c5eb9984f721c38d80fb5cbeae1298b9a2548f73, an attempt was made
to make the command interruptible, but this failed.

Although interruptible now, the command does not work anymore because
'exec' only executes the first command in the 'and' list and treats
everything behind it as this command's arguments.

So split the 'and' list into separate (and interruptible) commands.