Fix the command preparing files to be added to an archive
Preserve mode, ownership and timestamps of the files to be added, and
signal the end of options before the file list which is necessary to
disable further option processing so that files with names beginning
with a hyphen are not considered options.

Also, remove the space character between files and path, because the
list of files already ends with one.