Handle different info output formats of lrzip archives
Starting with lrzip version 0.650 the info output format changed
to a more human readable form.

(As part of this change, numbers were output in the local locale
and with thousands separators, which is planned to be reverted in
the next version of lrzip. Therefore, lrzip version 0.650 itself
is not supported.)