Improve drag operation and feedback in xa_dir_sidebar_drag_motion()
Allow and illustrate copy dropping on a TreeView row only. Change the
drag icon to illustrate that dropping is not possible in other parts
of dir_sidebar.

Also, clear the context data "tree_path" in other parts of dir_sidebar
so that dropping will not occur on the last touched TreeView row.