Completely rework the clipboard implementation
Since the clipboard data is not made available to any other process,
it is unnecessary to create an own GTK clipboard for it.

This also eliminates the need to convert the clipboard data to an
exportable format and allows direct access to the internally available
clipboard data storage.

Since xa_clipboard_clear() is now no longer a GtkClipboardClearFunc,
the unused parameter list can be omitted.