Allow editing the contents of a file compressed in an archive
Realize this via the clipboard with the new mode XA_CLIPBOARD_EDIT and
rephrase the condition for 'pasteable'.

Add an item 'Edit' to the popup menu, which extracts the file and offers
it for editing. A second item 'Replace' adds the extracted file back to
the archive (regardless of whether it was actually edited or not).

The edit function differs from the copy/cut functions of the clipboard
only in that xa_create_open_with_dialog() is called afterwards.

The replace function is the normal paste function of the clipboard and
replacement is only possible in the same archive where editing was done.

Each new addition to the clipboard deletes the previous contents. Only
one clipboard operation (copy/cut/paste or edit/replace) is possible at
a time.

This closes a requests by ousia, github issue #16.