Fix the generic column parser
The detection of filenames containing leading whitespace added with
commit 026c74cf9dbb9e69f03bc9b85995362c4e07b81b fails whenever the
assumption that there is a fixed column for the filenames is wrong.
It is more reasonable to assume a fixed spacing between filename and
the column before.

LAST_ITEM() is now rewritten for a one space character gap between
filename and the column before, i.e. LAST_ITEM() expects the line
pointer to already point to the filename when called. If the spacing
is larger, the argument to LAST_ITEM() simply needs to be incremented
accordingly afterwards.