Get rid of the term "rar5"
Since version 7 of the rar program, it is no longer possible to create
rar v4 type archives, so "rar5" can now be considered the standard and
should therefore be referred to only as "rar".

Instead, the format previously referred to as "rar", i.e. the rar v4
type format, is now referred to as (legacy) "rar4".