Refactor the recursive file collection
So far, for each compressor/archiver command line program, a list was
created that contained all files to be selected. If (recursively) all
subdirectories were to be included, this could result in argument lists
that were too long for the command line. For compressors/archivers with
an option to add recursively, this is not necessary and does not make
any sense.

So, use the respective option of the compressor/archiver command line
program if available, and refrain from custom recursion and file
collection when creating the file list.

For compressors/archivers without such option, however, the former
practice does not change.

This fixes part two of github issue #43, reported by Saijin-Naib.