Refactor the recursive file collection
So far, for each compressor/archiver command line program, a list was
created that contained all files to be selected. If (recursively) all
subdirectories were to be included, this could result in argument lists
that were too long for the command line. For compressors/archivers with
support to extract or delete recursively, this is not necessary and does
not make any sense.

So, use the appropriate directory pattern for the compressor/archiver
command line program, and refrain from custom recursion and file
collection when creating the file list.

For compressors/archivers without such support, however, the former
practice does not change.

This fixes part one of github issue #43, reported by mirh.