Make BSD sed happy
To edit a file in place without a backup, GNU sed would require either
no suffix argument or an empty argument appended to the option. BSD sed,
on the other hand, would require an empty argument to the option, but
separated by a space.

It is not possible to do both at the same time, so this is a workaround
using a temporary backup file.

This fixes github issue #186, reported by nunotexbsd.