Revise the design of the options page of the add dialog
Make the compression frame less high, increase the top padding of its
contents a bit, and decrease the top padding of the "No compression"
button.

This way the minimum height of the two bottom frames can be smaller.

Also, remove a call to gtk_button_set_alignment() which has no effect.