Replace xa_multi_extract_dialog_selection_changed()
It makes no sense to enable or disable the full_path option during
selection. Although this provides information about the capability of
the selected archive, the option will eventually apply to all added
archives.

Instead, count how many of the added archives do not support the option,
and enable the option if they all support it.

However, if the option is selected, it will be applied to all archives
that support it, even if it is disabled because some archives do not
support it.