Count the overwrite option in the same way as the full_path option
Count how many of the added archives do not support the option, and
enable the option if they all support it.

However, if the option is selected, it is going to be applied to all
archives that support it, even if it is disabled because some archives
do not support it.