Re-enable internal drag and drop
After commit eed981c32f7e3347b56c6d08b9a78a779698dc42, internal
drag and drop within an open archive or between open archives no
longer works if extended drag-and-drop support is disabled.

To make it work again, add the custom target xa-internal/uri-list,
which is always available and behaves identical to text/uri-list.

When a drop target requests the dragged data and the target type is
not XDirectSave, the type is always uri-list, so in this case set the
type of the selection data hard to text/uri-list. This will allow the
drop target's call to gtk_selection_data_get_uris() to also recognize
the xa-internal selection data.