Fix segmentation fault when dragging between open archives
When dragging from one open archive to another, the target archive is
the current one at drop time.

The selected files, however, must be collected from the source archive,
but because xa_concat_selected_filenames() automatically operates on the
current archive, it operates on the wrong one.

So replace the call to xa_concat_selected_filenames() with a call to
the new xa_concat_dragged_filenames(), which can be passed the required
archive to operate on.