lha: Fix handling of non-Unix lha archives
The only lha archives that have UID/GID metadata are those generated by
the Unix version of lha. Archives from other systems just show empty
whitespace at this point in the list output. As a result, trying to open
such archives with xarchiver just produced an empty window. This detects
when there is whitespace in the UID/GID column and handles it
appropriately.