lha: Recognize `-lhx-` compression algorithm
`-lhx-` is an extension of the `-lh7-` algorithm that can be generated
by `unlha32.dll` (and can be decompressed by Lhasa). Example archives
can be found in the Lhasa test suite:

<https://github.com/fragglet/lhasa/tree/master/test/archives/unlha32>

This change by itself is not sufficient to make xarchiver support these
archives; the parsing fixes in #214 are also required.