Fix handling of Mac OS lha archives
In Mac OS archives, "[Mac OS]" is listed in the permissions column,
which cannot be parsed correctly with a NEXT_ITEM macro.

Instead, copy the entire column and skip it afterwards - a method
that unfortunately depends on the length of the permissions column.

Reported by fragglet on github as part of his pull request #214.