Fix handling of missing timestamp information in lha archives
When the timestamp for a file in an lha archive is zero, empty
whitespace is displayed at this point in the list output.

Before grabbing and processing date and time, check whether the
STAMP column exists.

Reported by fragglet on github as part of his pull request #214,
but fixed differently.